// locadora-app/src/app/models/Veiculo/veiculo.model.ts
export interface Veiculo {
    id: string;
    locadoraId: string;
    nomeLocadoura: string;
    status: string;
    nomeMontadora: string;
    nomeModelo: string;
    numeroPortas: number;
    cor: string;
    fabricante: string;
    anoModelo: number;
    anoFabricacao: number;
    placa: string;
    chassi: string;
}